bugfix 0 mv pt for phantasmal units

+++ MoM/src/Combat.c
@@ -2522,7 +2522,8 @@ void Set_Movement_Cost_Map(int16_t battle_unit_idx)
             {
                 for(itr_x = 0; itr_x < COMBAT_GRID_WIDTH; itr_x++)
                 {
-                    if(battlefield->MoveCost_Sailing[((itr_y * COMBAT_GRID_WIDTH) + itr_x)] > battlefield->MoveCost_Ground[((itr_y * COMBAT_GRID_WIDTH) + itr_x)])
+                    /* cast to uint8_t so INF (stored as -1 in int8_t) reads as 255 = worst-cost; signed compare treated INF as smaller than any real cost and produced INF for every cell, making UA_NONCORPOREAL / Wraith-Form / swimming units immobile. */
+                    if((uint8_t)battlefield->MoveCost_Sailing[((itr_y * COMBAT_GRID_WIDTH) + itr_x)] > (uint8_t)battlefield->MoveCost_Ground[((itr_y * COMBAT_GRID_WIDTH) + itr_x)])
                     {
                         _cmbt_movepath_cost_map[((itr_y * COMBAT_GRID_WIDTH) + itr_x)] = battlefield->MoveCost_Ground[((itr_y * COMBAT_GRID_WIDTH) + itr_x)];
                     }
